% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{irregular_list_to_dataframe}
\alias{irregular_list_to_dataframe}
\title{Convert irregular list into two column dataframe}
\usage{
irregular_list_to_dataframe(irregular_list, col_names = c("keyword", "vector"))
}
\arguments{
\item{irregular_list}{list whose lengths are of variable length, making it difficult to convert into dataframe}

\item{col_names}{vector of names to use for newly created two column dataframe}
}
\description{
This function will convert an irregular list into a two column dataframe, where the first column is the list keyword and the second column is the vector within the keyword
}
\examples{
irregular_list_to_dataframe()
}
\keyword{utility}
