% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq_functions.R
\name{plotPCAmulti}
\alias{plotPCAmulti}
\title{DESeq plotPCA for all PCs}
\usage{
plotPCAmulti(object, intgroup = "condition", ntop = 500)
}
\arguments{
\item{object}{a DESeqTransform object, with data in assay(x), produced for example by either rlog or varianceStabilizingTransformation.}

\item{intgroup}{interesting groups: a character vector of names in colData(x) to use for grouping}

\item{ntop}{number of top genes to use for principal components, selected by highest row variance}
}
\description{
This function expands on the plotPCA function of DESeq2 that only returns the data for the first two principal components. This function will all PCs
}
\examples{
plotPCAmulti()
}
\keyword{DESeq2}
