% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_functions.R
\name{SC_load_CellPlex}
\alias{SC_load_CellPlex}
\title{Load multiple 10x CellPlex output in Seurat object list}
\usage{
SC_load_CellPlex(
  sample_info,
  sample_name_column,
  sample_location_column,
  metadata_columns
)
}
\arguments{
\item{sample_info}{A sample information datafram containing; sample names, paths to cellranger multi outputs, and as many parameter columns as wanted to add to metadata}

\item{sample_name_column}{Name of column that contains the sample names}

\item{sample_location_column}{Name of column that contains the path to the cellranger multi outputs}

\item{metadata_columns}{Vector containing all the names of the columns that need to be added as meta data}
}
\value{
Returns a list of Seurat objects to be used for downstream integration
}
\description{
This function takes information from a sample info dataframe containing the sample names, location of cellranger multi outputs, and metadata columns and loads each sample in as a Seurat object.
}
\examples{
plotPCAmulti()
}
\keyword{Seurat}
