% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq_functions.R
\name{summarize_plotPCAmulti}
\alias{summarize_plotPCAmulti}
\title{Summarize all PCs from plotPCAmulti}
\usage{
summarize_plotPCAmulti(data, PCs = Inf)
}
\arguments{
\item{data}{Output dataframe from plotPCAmulti}

\item{PCs}{Number of principal components to include (default: All)}
}
\description{
Function takes the plotPCAmulti dataframe as input and will show all PCs together.
}
\examples{
summarize_plotPCAmulti()
}
\keyword{DESeq2}
